+++ src/map/map-polygon-methods.js
@@ -10,6 +10,8 @@
  * @return {array} Array of points.
  */
 export function getPoints(ray = false) {
+    if(!this.active)
+        return [];
     return this._points;
 };
 
@@ -23,6 +25,8 @@ export function getPoints(ray = false) {
  * @return {array} Array of Phaser.Geom.Line objects.
  */
 export function getSegments() {
+    if(!this.active)
+        return [];
     return this._segments;
 };
 
@@ -36,6 +40,9 @@ export function getSegments() {
  * @return {object} Map object.
  */
 export function updateMap() {
+    if(!this.active)
+        return this;
+
     let points = [];
     let segments = [];